Removed bugs and added test program
Rev3 replaces Rev2. Several bugs were fixed.
Number of leds per signal and startpin is now calculated by the program

Added a testprogram to test the decoder and its functions. Instead of
connecting to DCC, the configured addresses can be entered with the
monitor program. Separated by a comma the aspect (0 or 1) is entered.
The test program will execute this as if the dress and aspect was
received over DCC.